register page refactor
