+++ front-end/src/Views/RegisterPage/components/Company/companyDataForm.js
@@ -1,20 +1,21 @@
-import React from "react";
+import React from 'react';
+import { onChange } from '../../functions/handlers';
 import {Card, Form} from "react-bootstrap";
-import {onChange} from "../functions/handlers";
 
-class HomeDataForm extends React.Component {
+class CompanyDataForm extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
             name_of_place: "",
             street: "",
             city: "",
-            city_code: ""
+            city_code: "",
+            company_nip: ""
         }
     }
 
-    render() {
-        const { name_of_place, street, city, city_code } = this.state;
+    render () {
+        let { name_of_place, street, city, city_code, company_nip } = this.state;
         let {data, onBlur} = this.props;
         return (
             <Card
@@ -22,28 +23,27 @@ class HomeDataForm extends React.Component {
                 className="loginPage__cardSection col-lg mr-lg-3"
             >
                 <Card.Header as="h4" className="">
-                    Dane placówki
+                    Dane firmy
                 </Card.Header>
-                <Card.Body className="account_form_card">
-                    {/*console.log(this.state)*/}
+                <Card.Body className="">
                     <Form.Group controlId="formGroupNameOfPlace">
                         <Form.Control
                             name="name_of_place"
-                            type="tel"
+                            type="text"
                             autoComplete="on"
-                            placeholder="Nazwa placówki"
+                            placeholder="Nazwa firmy"
                             defaultValue={name_of_place}
                             onBlur={e => onChange(onBlur, data, e)}
                             required
                         />
                         <Form.Control.Feedback type="invalid">
-                            Podaj nazwę placówki
+                            Podaj nazwę firmy
                         </Form.Control.Feedback>
                     </Form.Group>
                     <Form.Group controlId="formGroupStreet">
                         <Form.Control
                             name="street"
-                            type="tel"
+                            type="text"
                             autoComplete="on"
                             placeholder="Ulica"
                             defaultValue={street}
@@ -51,14 +51,13 @@ class HomeDataForm extends React.Component {
                             required
                         />
                         <Form.Control.Feedback type="invalid">
-                            Podaj ulicę na której znajduje się placówka
+                            Podaj ulicę na której znajduje się firma
                         </Form.Control.Feedback>
                     </Form.Group>
-                    {/* <section className="row"> */}
                     <Form.Group controlId="formGroupCity" className="">
                         <Form.Control
                             name="city"
-                            type="tel"
+                            type="text"
                             autoComplete="on"
                             placeholder="Nazwa miasta"
                             defaultValue={city}
@@ -73,7 +72,7 @@ class HomeDataForm extends React.Component {
                     <Form.Group controlId="formGroupCityCode" className="">
                         <Form.Control
                             name="city_code"
-                            type="tel"
+                            type="text"
                             autoComplete="on"
                             placeholder="Kod pocztowy"
                             defaultValue={city_code}
@@ -84,11 +83,24 @@ class HomeDataForm extends React.Component {
                             Podaj kod pocztowy
                         </Form.Control.Feedback>
                     </Form.Group>
-                    {/* </section> */}
+                    <Form.Group controlId="formGroupCompanyNip" className="">
+                        <Form.Control
+                            name="company_nip"
+                            type="text"
+                            autoComplete="on"
+                            placeholder="NIP"
+                            defaultValue={company_nip}
+                            onBlur={e => onChange(onBlur, data, e)}
+                            required
+                        />
+                        <Form.Control.Feedback type="invalid">
+                            Podaj NIP
+                        </Form.Control.Feedback>
+                    </Form.Group>
                 </Card.Body>
             </Card>
-        )
+        );
     }
 }
 
-export default HomeDataForm;
+export default CompanyDataForm;
