+++ front-end/src/Views/RegisterPage/components/FosterHome/homeDataForm.js
@@ -1,21 +1,20 @@
-import React from 'react';
-import { onChange } from '../functions/handlers';
+import React from "react";
 import {Card, Form} from "react-bootstrap";
+import {onChange} from "../../functions/handlers";
 
-class CompanyDataForm extends React.Component {
+class HomeDataForm extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
             name_of_place: "",
             street: "",
             city: "",
-            city_code: "",
-            company_nip: ""
+            city_code: ""
         }
     }
 
-    render () {
-        let { name_of_place, street, city, city_code, company_nip } = this.state;
+    render() {
+        const { name_of_place, street, city, city_code } = this.state;
         let {data, onBlur} = this.props;
         return (
             <Card
@@ -23,27 +22,28 @@ class CompanyDataForm extends React.Component {
                 className="loginPage__cardSection col-lg mr-lg-3"
             >
                 <Card.Header as="h4" className="">
-                    Dane firmy
+                    Dane placówki
                 </Card.Header>
-                <Card.Body className="">
+                <Card.Body className="account_form_card">
+                    {/*console.log(this.state)*/}
                     <Form.Group controlId="formGroupNameOfPlace">
                         <Form.Control
                             name="name_of_place"
-                            type="text"
+                            type="tel"
                             autoComplete="on"
-                            placeholder="Nazwa firmy"
+                            placeholder="Nazwa placówki"
                             defaultValue={name_of_place}
                             onBlur={e => onChange(onBlur, data, e)}
                             required
                         />
                         <Form.Control.Feedback type="invalid">
-                            Podaj nazwę firmy
+                            Podaj nazwę placówki
                         </Form.Control.Feedback>
                     </Form.Group>
                     <Form.Group controlId="formGroupStreet">
                         <Form.Control
                             name="street"
-                            type="text"
+                            type="tel"
                             autoComplete="on"
                             placeholder="Ulica"
                             defaultValue={street}
@@ -51,13 +51,14 @@ class CompanyDataForm extends React.Component {
                             required
                         />
                         <Form.Control.Feedback type="invalid">
-                            Podaj ulicę na której znajduje się firma
+                            Podaj ulicę na której znajduje się placówka
                         </Form.Control.Feedback>
                     </Form.Group>
+                    {/* <section className="row"> */}
                     <Form.Group controlId="formGroupCity" className="">
                         <Form.Control
                             name="city"
-                            type="text"
+                            type="tel"
                             autoComplete="on"
                             placeholder="Nazwa miasta"
                             defaultValue={city}
@@ -72,7 +73,7 @@ class CompanyDataForm extends React.Component {
                     <Form.Group controlId="formGroupCityCode" className="">
                         <Form.Control
                             name="city_code"
-                            type="text"
+                            type="tel"
                             autoComplete="on"
                             placeholder="Kod pocztowy"
                             defaultValue={city_code}
@@ -83,24 +84,11 @@ class CompanyDataForm extends React.Component {
                             Podaj kod pocztowy
                         </Form.Control.Feedback>
                     </Form.Group>
-                    <Form.Group controlId="formGroupCompanyNip" className="">
-                        <Form.Control
-                            name="company_nip"
-                            type="text"
-                            autoComplete="on"
-                            placeholder="NIP"
-                            defaultValue={company_nip}
-                            onBlur={e => onChange(onBlur, data, e)}
-                            required
-                        />
-                        <Form.Control.Feedback type="invalid">
-                            Podaj NIP
-                        </Form.Control.Feedback>
-                    </Form.Group>
+                    {/* </section> */}
                 </Card.Body>
             </Card>
-        );
+        )
     }
 }
 
-export default CompanyDataForm;
+export default HomeDataForm;
